feat: 68 - move result list to results_view_model
